Rename local utils to app_utils. Refactor
